Refactor card play logic into separate modules
